lab5_few_shot/pyproject.toml
[project]
name = "llm-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.128.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
